requirements.txt
customtkinter
pystray
Pillow
pypresence
paho-mqtt
playwright
beautifulsoup4
httpx
requests